refactor: separate data responses from control responses
