+++ src/momento/responses/__init__.py
@@ -8,66 +8,82 @@ from .control.signing_key.list import (
     SigningKey,
 )
 from .control.signing_key.revoke import RevokeSigningKey, RevokeSigningKeyResponse
-from .dictionary_data import (
+from .data.dictionary_data.dictionary_fetch import (
     CacheDictionaryFetch,
     CacheDictionaryFetchResponse,
+)
+from .data.dictionary_data.dictionary_get_field import (
     CacheDictionaryGetField,
     CacheDictionaryGetFieldResponse,
+)
+from .data.dictionary_data.dictionary_get_fields import (
     CacheDictionaryGetFields,
     CacheDictionaryGetFieldsResponse,
+)
+from .data.dictionary_data.dictionary_increment import (
     CacheDictionaryIncrement,
     CacheDictionaryIncrementResponse,
+)
+from .data.dictionary_data.dictionary_remove_field import (
     CacheDictionaryRemoveField,
     CacheDictionaryRemoveFieldResponse,
+)
+from .data.dictionary_data.dictionary_remove_fields import (
     CacheDictionaryRemoveFields,
     CacheDictionaryRemoveFieldsResponse,
+)
+from .data.dictionary_data.dictionary_set_field import (
     CacheDictionarySetField,
     CacheDictionarySetFieldResponse,
+)
+from .data.dictionary_data.dictionary_set_fields import (
     CacheDictionarySetFields,
     CacheDictionarySetFieldsResponse,
 )
-from .list_data import (
+from .data.list_data.list_concatenate_back import (
     CacheListConcatenateBack,
     CacheListConcatenateBackResponse,
+)
+from .data.list_data.list_concatenate_front import (
     CacheListConcatenateFront,
     CacheListConcatenateFrontResponse,
-    CacheListFetch,
-    CacheListFetchResponse,
-    CacheListLength,
-    CacheListLengthResponse,
-    CacheListPopBack,
-    CacheListPopBackResponse,
-    CacheListPopFront,
-    CacheListPopFrontResponse,
-    CacheListPushBack,
-    CacheListPushBackResponse,
+)
+from .data.list_data.list_fetch import CacheListFetch, CacheListFetchResponse
+from .data.list_data.list_length import CacheListLength, CacheListLengthResponse
+from .data.list_data.list_pop_back import CacheListPopBack, CacheListPopBackResponse
+from .data.list_data.list_pop_front import CacheListPopFront, CacheListPopFrontResponse
+from .data.list_data.list_push_back import CacheListPushBack, CacheListPushBackResponse
+from .data.list_data.list_push_front import (
     CacheListPushFront,
     CacheListPushFrontResponse,
+)
+from .data.list_data.list_remove_value import (
     CacheListRemoveValue,
     CacheListRemoveValueResponse,
 )
-from .response import CacheResponse, ControlResponse
-from .scalar_data import (
-    CacheDelete,
-    CacheDeleteResponse,
-    CacheGet,
-    CacheGetResponse,
-    CacheIncrement,
-    CacheIncrementResponse,
-    CacheSet,
+from .data.scalar_data.delete import CacheDelete, CacheDeleteResponse
+from .data.scalar_data.get import CacheGet, CacheGetResponse
+from .data.scalar_data.increment import CacheIncrement, CacheIncrementResponse
+from .data.scalar_data.set import CacheSet, CacheSetResponse
+from .data.scalar_data.set_if_not_exists import (
     CacheSetIfNotExists,
     CacheSetIfNotExistsResponse,
-    CacheSetResponse,
 )
-from .set_data import (
+from .data.set_data.set_add_element import (
     CacheSetAddElement,
     CacheSetAddElementResponse,
+)
+from .data.set_data.set_add_elements import (
     CacheSetAddElements,
     CacheSetAddElementsResponse,
-    CacheSetFetch,
-    CacheSetFetchResponse,
+)
+from .data.set_data.set_fetch import CacheSetFetch, CacheSetFetchResponse
+from .data.set_data.set_remove_element import (
     CacheSetRemoveElement,
     CacheSetRemoveElementResponse,
+)
+from .data.set_data.set_remove_elements import (
     CacheSetRemoveElements,
     CacheSetRemoveElementsResponse,
 )
+from .response import CacheResponse, ControlResponse
