+++ src/momento/responses/data/dictionary_data/dictionary_set_fields.py
@@ -1,0 +1,28 @@
+from abc import ABC
+
+from ...mixins import ErrorResponseMixin
+from ...response import CacheResponse
+
+
+class CacheDictionarySetFieldsResponse(CacheResponse):
+    """Parent response type for a cache `dictionary_set_fields` request. Its subtypes are:
+
+    - `CacheDictionarySetFields.Success`
+    - `CacheDictionarySetFields.Error`
+
+    See `SimpleCacheClient` for how to work with responses.
+    """
+
+
+class CacheDictionarySetFields(ABC):
+    """Groups all `CacheDictionarySetFieldsResponse` derived types under a common namespace."""
+
+    class Success(CacheDictionarySetFieldsResponse):
+        """Indicates the request was successful."""
+
+    class Error(CacheDictionarySetFieldsResponse, ErrorResponseMixin):
+        """Contains information about an error returned from a request:
+
+        - `error_code`: `MomentoErrorCode` value for the error.
+        - `messsage`: a detailed error message.
+        """
