+++ src/momento/responses/data/set_data/set_add_elements.py
@@ -1,0 +1,28 @@
+from abc import ABC
+
+from ...mixins import ErrorResponseMixin
+from ...response import CacheResponse
+
+
+class CacheSetAddElementsResponse(CacheResponse):
+    """Parent response type for a `set_add_elements` request. Its subtypes are:
+
+    - `CacheSetAddElements.Success`
+    - `CacheSetAddElements.Error`
+
+    See `SimpleCacheClient` for how to work with responses.
+    """
+
+
+class CacheSetAddElements(ABC):
+    """Groups all `CacheSetAddElementsResponse` derived types under a common namespace."""
+
+    class Success(CacheSetAddElementsResponse):
+        """Indicates the elements were added."""
+
+    class Error(CacheSetAddElementsResponse, ErrorResponseMixin):
+        """Contains information about an error returned from a request:
+
+        - `error_code`: `MomentoErrorCode` value for the error.
+        - `messsage`: a detailed error message.
+        """
