--- src/momento/responses/dictionary_data/dictionary_remove_field.py
@@ -1,28 +1,0 @@
-from abc import ABC
-
-from ..mixins import ErrorResponseMixin
-from ..response import CacheResponse
-
-
-class CacheDictionaryRemoveFieldResponse(CacheResponse):
-    """Parent response type for a cache `dictionary_remove_field` request. Its subtypes are:
-
-    - `CacheDictionaryRemoveField.Success`
-    - `CacheDictionaryRemoveField.Error`
-
-    See `SimpleCacheClient` for how to work with responses.
-    """
-
-
-class CacheDictionaryRemoveField(ABC):
-    """Groups all `CacheDictionaryRemoveFieldResponse` derived types under a common namespace."""
-
-    class Success(CacheDictionaryRemoveFieldResponse):
-        """Indicates the request was successful."""
-
-    class Error(CacheDictionaryRemoveFieldResponse, ErrorResponseMixin):
-        """Contains information about an error returned from a request:
-
-        - `error_code`: `MomentoErrorCode` value for the error.
-        - `messsage`: a detailed error message.
-        """
