--- src/momento/responses/list_data/list_remove_value.py
@@ -1,28 +1,0 @@
-from abc import ABC
-
-from ..mixins import ErrorResponseMixin
-from ..response import CacheResponse
-
-
-class CacheListRemoveValueResponse(CacheResponse):
-    """Response type for a `list_remove_value` request. Its subtypes are:
-
-    - `CacheListRemoveValue.Success`
-    - `CacheListRemoveValue.Error`
-
-    See `SimpleCacheClient` for how to work with responses.
-    """
-
-
-class CacheListRemoveValue(ABC):
-    """Groups all `CacheListRemoveValueResponse` derived types under a common namespace."""
-
-    class Success(CacheListRemoveValueResponse):
-        """Indicates removing the values was successful."""
-
-    class Error(CacheListRemoveValueResponse, ErrorResponseMixin):
-        """Indicates an error occured in the request:
-
-        - `error_code`: `MomentoErrorCode` value for the error.
-        - `messsage`: a detailed error message.
-        """
